+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changes
 - Fix some bugs in buttsbot
 - Add command line option to save file with a command list
+- Create command list